% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used_pkgs.R
\name{get_used_pkg}
\alias{get_used_pkg}
\title{Get Loaded Packages}
\usage{
get_used_pkg(stored_file)
}
\arguments{
\item{stored_file}{file stored as a vector of character.}
}
\description{
Get Loaded Packages
}

