% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_R_files.R
\name{list_R_files}
\alias{list_R_files}
\title{List all R files in directory}
\usage{
list_R_files(path = ".")
}
\arguments{
\item{path}{A character string indicating a path or a list of paths}
}
\value{
A character vector of found files with extensions in ".R"
}
\description{
List all R files in directory
}

